fix: replace sidepanel component with frappe-ui Dialog
